+++ CHANGELOG.md
@@ -1,11 +1,18 @@
 # Change Log
 
-All notable changes to the **Secure Notes** extension will be documented in this file.
+All notable changes to the **VSC Secure Notes** extension will be documented in this file.
+
+## v1.1.0 - 2026-03-28
+
+### Added
+
+- Settings to set the AES Options (mode and keysize).
+- Encryption module now adapted to support CBC and CTR as well.
 
 ## v1.0.0 - 2026-03-28
 
 ### Added
 
 - Initial release — VS Code port of the [Joplin Secure Notes plugin](https://github.com/cipherswami/joplin-plugin-secure-notes)
-- `Secure Notes: Encrypt` command to encrypt open file or selected text with a password
-- `Secure Notes: Decrypt` command to decrypt previously encrypted notes
+- `VSC Secure Notes: Encrypt` command to encrypt open file or selected text with a password
+- `VSC Secure Notes: Decrypt` command to decrypt previously encrypted notes